refactor: generalize evm call script describing hook
