+++ src/apps/voting/app-logic.js
@@ -1,6 +1,5 @@
 import { useCallback, useMemo } from 'react'
-import { usePath } from '@/hooks/shared'
-import { useApi } from './hooks/shared'
+import { useApi, usePath } from '@/hooks/shared'
 import { useAppState } from './providers/VotingProvider'
 import { EMPTY_CALLSCRIPT } from './evmscript-utils'
 import usePanelState from './hooks/usePanelState'
