+++ src/hooks/shared/useApi.js
@@ -1,5 +1,6 @@
 export default function useApi() {
   return {
+    requestAddressIdentityModification: () => '',
     resolveAddressIdentity: async () => '',
     external: async () => '',
     call: async () => '',
